changed plotting routine to use original magnetogram rather than processed
To get rid of the grey background off limb, i stopped using the processed magnetogram in the plotting routine, and instead use the original inmap. So this is changed in the wrapper and in the plot code.

In the plot code I added a flip if magetogram the wrong way around (which came from the processing code without all the extra stuff).

Also switched on saving fits files for the craic (and science)